% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{clColours}
\alias{clColours}
\title{Default colours for analyses}
\format{An object of class \code{character} of length 8.}
\usage{
clColours
}
\description{
Default colours for analyses
}
\keyword{datasets}
