% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{clInitializeTernarySplits}
\alias{clInitializeTernarySplits}
\title{Initialize ternary plots for partition plotting}
\usage{
clInitializeTernarySplits(fontSize = 1, xlim = NULL, ylim = NULL,
  gridCol = rgb(0.92, 0.92, 0.92), padding = 0.1)
}
\description{
Initialize ternary plots for partition plotting
}
