% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{clPhyDat}
\alias{clPhyDat}
\alias{clMatrices}
\title{100 simulated data matrices}
\format{\itemize{
\item \code{clPhyDat}: A list with 100 entries, each comprising a phyDat object
of 55 characters for 22 taxa.
\item \code{clMatrices}: A list with 100 entries, each comprising a list of character
tokens for each simulated character, as read from raw nexus files using
\code{ape::read.nexus.data}.  The four dummy 'characters' have been removed.
}}
\source{
\url{http://datadryad.org/resource/doi:10.5061/dryad.7dq0j}
}
\usage{
clPhyDat

clMatrices
}
\description{
Contains the 100 simulated matrices generated by Congreve & Lamsdell (2016) using a
heterogeneous Markov-k model, generated from the \link{clReferenceTree} topology, with
all branches sharing an equal length.
}
\references{
\insertRef{Congreve2016}{Quartet}
\insertRef{Congreve2016dd}{Quartet}
}
\keyword{datasets}
