% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{clReferenceTree}
\alias{clReferenceTree}
\title{Tree topology for matrix simulation}
\format{A single phylogenetic tree saved as an object of class \code{phylo}}
\source{
\insertRef{Congreve2016}{Quartet}
}
\usage{
clReferenceTree
}
\description{
The tree topology used to generate the matrices in \code{\link{clMatrices}}
Congreve & Lamsdell (2016)
}
\examples{
  library(ape) # Contains tree plotting functions
  data(clReferenceTree)
  plot(clReferenceTree)

}
\references{
\insertRef{Congreve2016}{Quartet}
\insertRef{Congreve2016dd}{Quartet}
}
\keyword{datasets}
