% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/points.R
\name{SplitsToPoints}
\alias{SplitsToPoints}
\title{Convert splits data to ternary-plottable points}
\usage{
SplitsToPoints(itemData)
}
\arguments{
\item{itemData}{A line from \code{clBootGcPartitions} or equivalent}
}
\value{
Number of partitions resolved in tree 2 only, different, and same in both
trees, numbered ready to appear on a ternary plot.
}
\description{
Convert splits data to ternary-plottable points
}
