% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{clInitializeTernaryQuarts}
\alias{clInitializeTernaryQuarts}
\title{Initialize ternary plots for quartet plotting}
\usage{
clInitializeTernaryQuarts(zoom = 1, padding = 0.1, gridLines = 10,
  fontSize = 1, gridCol = rgb(0.92, 0.92, 0.92),
  backgroundCol = rgb(0.985, 0.985, 0.992), xLim = c(0, 1/zoom) - 0.01,
  yLim = c(0.5 - (1/zoom), 0.5))
}
\description{
Initialize ternary plots for quartet plotting
}
