% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/points.R
\name{clPlotQuartets}
\alias{clPlotQuartets}
\title{Plot results}
\usage{
clPlotQuartets(dataset, tree, cex = 1.1, pch = 2, COL = clColours,
  ...)
}
\arguments{
\item{tree}{Integer specifying which tree to plot}

\item{cex, pch, \dots}{Graphical parameters to pass to \link[Ternary:JoinTheDots]{Ternary::JoinTheDots}}

\item{COL}{Named vector specifying colour in which to plot each analysis}
}
\description{
Plot results
}
\author{
Martin R. Smith
}
