% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{clQuartets}
\alias{clQuartets}
\title{Congreve and Lamsdell quartet distances}
\format{An object of class \code{list} of length 8.}
\source{
Congreve, C. R. & Lamsdell, J. C. (2016). Implied weighting and its
utility in palaeontological datasets: a study using modelled phylogenetic
matrices. Palaeontology 59(3), 447--465. \url{doi:10.1111/pala.12236}.
}
\usage{
clQuartets
}
\description{
Distance of CL trees from generative tree according to the quartet distance.
}
\details{
For each of the 100 matrices generated by Congreve & Lamsdell (2016), I conducted
phylogenetic analysis under different methods:

\itemize{
\item{\code{markov}: }{Using the Markov K model in MrBayes.}
\item{\code{equal}: }{Using equal weights in TNT.}
\item{\code{implied1}, \code{implied2}, \code{implied3}, \code{implied5}, \code{implied10}: }{Using implied weights in TNT,
with the concavity constant (\emph{k}) set to 1, 2, 3, 5, or 10}
\item{\code{impliedC}: }{By taking the strict \emph{c}onsensus of all trees recovered by implied
weights parsimony analysis under the \emph{k} values 2, 3, 5 and 10 (but not 1).}
}

For each analysis, I recorded the strict consensus of all optimal trees, and also
the consensus of trees that were suboptimal by a specified degree.

I then calculated, of the total number of quartets that were resolved in the
reference tree, how many were the \emph{s}ame or \emph{d}ifferent in the tree that resulted from
the phylogenetic analysis, and how many were not resolved in this tree (\emph{r2}).

The data object contains a list whose elements are named after the methods, as listed above.

Each list entry is a three-dimensional array, whose dimensions are:
\enumerate{
\item The suboptimality of the tree: for \emph{markov}, the consensus at a 50%,
52.5%, ... 97.5%, 100% posterior probability; for \emph{equal}, the consensus of
all trees that are 0, 1, .... 19, 20 steps less optimal than the optimal
tree; for \emph{implied}, the consensus of all trees that are 0.73^(19:0)
less optimal than the optimal tree.
\item The number of quartets in total, the same, different, resolved
in the estimated tree but not the generative tree (= 0), resolved in the
generative tree but not the estimated tree
\item The number of the matrix, from 1 to 100.
}
}
\seealso{
\link{clMatrices}, \link{clReferenceTree}.
}
\keyword{datasets}
