% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{clCI}
\alias{clCI}
\title{Consistency indices}
\format{An object of class \code{numeric} of length 100.}
\usage{
clCI
}
\description{
Consistency indices of Congreve & Lamsdell datasets.
}
\keyword{datasets}
